# perf(data_chunk): add benchmark for `DataChunk::compact()`