--- a/src/common/Cargo.toml
+++ b/src/common/Cargo.toml
@@ -127,6 +127,10 @@ harness = false
 name = "bench_data_chunk_encoding"
 harness = false
 
+[[bench]]
+name = "bench_data_chunk_compact"
+harness = false
+
 [[bin]]
 name = "example-config"
 path = "src/bin/default_config.rs"